Use secure URLs when connecting to wordpress.org
Since .org is fully SSL/TLS, not using 'https' results in a 301 redirect
with every request. Switching speeds up the requests by hitting the
correct endpoint, and provides an extra level of security.
Most of the URLs can be switched by adding 's'; the plugins SVN URL will
respond on a variety of domain combinations, but the security cert is
issued to protect '*.svn.wordpress.org', so it needed to be completely
changed here to avoid certificate errors.

Conflicts:
	update